requirements.txt
neptune-client
numpy
torch
pyyaml
